Share new space keys with FI Wide layout
